+++ src/pages/layout/Layout.tsx
@@ -1,4 +1,5 @@
 import { ReactNode, useEffect, useRef } from 'react'
+import Footer from '../../components/shared/footer'
 
 type LayoutProps = {
   components: ReactNode
@@ -8,6 +9,7 @@ export default function Layout({ components }: LayoutProps) {
   return (
     <>
       <div>{components}</div>
+      <Footer />
     </>
   )
 }